// src/app/models/common.models.ts
export enum DataType {
  PositiveNegative = 1,
  Numeric = 2,
  Text = 3,
  Percentage = 4,
  Combobox = 5,
}

export enum Status {
  Inactive = 0,
  Active = 1,
  Deleted = 2,
}

export enum MedicalNoteTypeEnum {
  Initial = 1,
  Admission = 2,
  Referral = 3,
  Interconsultation = 4,
  Evolution = 5,
  Discharge = 6,
}